EEPROM Programmer: New circuit, with more flexibility
Redesigned the circuit to be able to control the OE, CS and WE pins from the
EEPROM, and the OE, SRCLR pins from the shift register.

This gives us more total control of all the pins in both chips, and
enables us to write to the EEPROM (before the WE pin was always HIGH).

I also added a status LED for monitoring.